update event details
