+++ app/src/main/java/com/example/marvelverse/app/ui/events/details/EventDetailsViewModel.kt
@@ -1,8 +1,11 @@
-package com.example.marvelverse.app.ui.events
+package com.example.marvelverse.app.ui.events.details
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
+import com.example.marvelverse.app.ui.home.interfaces.CharacterInteractionListener
+import com.example.marvelverse.app.ui.home.interfaces.ComicInteractionListener
+import com.example.marvelverse.app.ui.home.interfaces.SeriesInteractionListener
 import com.example.marvelverse.data.repositories.MarvelRepository
 import com.example.marvelverse.domain.entities.main.Comic
 import com.example.marvelverse.domain.entities.main.Series
@@ -11,18 +14,19 @@ import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers
 import io.reactivex.rxjava3.disposables.CompositeDisposable
 import io.reactivex.rxjava3.schedulers.Schedulers
 
-class EventViewModel : ViewModel() {
+class EventDetailsViewModel : ViewModel(), CharacterInteractionListener, ComicInteractionListener,
+	SeriesInteractionListener {
 
 	private val compositeDisposable by lazy { CompositeDisposable() }
-	private val _character: MutableLiveData<List<Character>?> = MutableLiveData()
-	private val _comics: MutableLiveData<List<Comic>?> = MutableLiveData()
-	private val _series: MutableLiveData<List<Series>?> = MutableLiveData()
-	val comics: LiveData<List<Comic>?>
-		get() = _comics
-	val series: LiveData<List<Series>?>
-		get() = _series
+	private val _character: MutableLiveData<List<Character>> = MutableLiveData()
+	private val _comics: MutableLiveData<List<Comic>> = MutableLiveData()
+	private val _series: MutableLiveData<List<Series>> = MutableLiveData()
 
-	val character: LiveData<List<Character>?>
+	val comics: LiveData<List<Comic>>
+		get() = _comics
+	val series: LiveData<List<Series>>
+		get() = _series
+	val character: LiveData<List<Character>>
 		get() = _character
 
 	fun getCharacters(url: String) {
@@ -60,5 +64,14 @@ class EventViewModel : ViewModel() {
 
 	override fun onCleared() {
 		compositeDisposable.clear()
+	}
+
+	override fun onCharacterClick(character: Character) {
+	}
+
+	override fun onComicClick(comic: Comic) {
+	}
+
+	override fun onSeriesClick(series: Series) {
 	}
 }
